Fix unescaping of character sequences when it should not
The HGFS server when reading directory entries checks for
an escape character sequence and if found unescapes it before
sending the entry to the HGFS client. However, if the sequence
happens to be one generated by a user legitimately on the host
and not because the HGFS server found an illegal character, the
unescape mechanism has no way to tell and currently does more
undo operations than the escape. This leads to some file
names being mistranslated and therefore inaccessible e.g.

Here is an overview of the current scheme as best I could tell.

The escape character is the percent.
The character to escape the escape character is the right square
parenthesis.

Escape looks for the following characters or sequences:
platform illegal char -> platform substitute char and '%'
']%' - > ']]%'
platform substitute char and '%' -> substitute char and ']%'
Windows reserved names or ending with a period

Currently the unescaping code checks for the ']%' character
sequence resulting from item 2 or 3 above is preceded by a
character to replace an illegal one from a client. The code
simply assumes the preceding character is of the right type,
and so modifies the string.

A file "@]%.txt" on the host is unescaped in a directory read to
the client as "@%.txt" where the client will escape it again if
required. When a client application tries to open the file, the
HGFS client returns the "@%.txt" to the host HGFS server. There
the escape mechanism for Linux/OSX does not see this as a
sequence requiring escaping as the '@' is not a substitute
character. This means the file name is not the same as the original
on the host and it either fails to open or the server will create
a new file with the modified name.

The escape undo code now only strips out the escape-escape character
if it is preceded by the same character or a platform substitute
character (i.e. a sequence that the escape mechanism generated).

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>