vmtoolsd: separate loading and initializing plugins.
Currently the code that loads the plugins also initializes them right
away. This is undesired when using appLoader, since in the process of
figuring out dependencies, it may fork the process again and cause
the same plugin to be initialized multiple times.

This doesn't cause any issues other than a few (one time) memory leaks
and wasted work, but it's better to avoid this.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>