Unity/X11: Clear tools.capability.unity in response to X I/O errors.
Summary:
desktopEvents:
- No longer reloading self in X I/O error handler.
- Emitting GLib signal prior to shutting down RPC channel.
unity:
- Attach to new XIOError signal a callback which clears the Unity
capability.

Long version:
In response to an X I/O error, cleanup of the vmtoolsd container is
kinda hard.  From the plugin providing the X I/O error handler, we
can't directly call the cleanup routines, and even if we could,
there's nothing to prevent any of them from attempting additional
X calls during cleanup, bringing us back to square 1.

For the most part, this lameness hasn't caused any serious problems.
One annoyance, however, is that if our user was in Unity mode at the
time, vmusr's untimely termination wouldn't inform the host UI that
vmusr was on its way out, leaving the host UI believing it should
remain in Unity mode.

While we can't perform a full, correct cleanup, we can at least undo
a few small critical things, such as saying, "Hey, host UI.  Please
to be doing the needful and bumping us out of Unity now, kthx."  This
is accomplished by a) emitting a GLib signal in the XIO error handler,
and b) connecting that to a Unity plugin callback which clears the
capability.

Brownie points: I got rid of the dumb call to Reload_Do that I was
responsible for years ago.  There really isn't a good reason for an
X client to restart itself in response to an X I/O error, so let's
just stop, yeah?

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>