File_MakeTempEx2 race
There is a nasty race when two threads/processes call
File_MakeTempEx2. It is possible for one caller to be creating
files while the other is creating directories. Since there is
one "name space" there can be collisions that are a pain to deal
with.

The solution is simple - files are created with odd numbers,
directories with even numbers. This way the available exclusions
(via O_EXCL and mkdir) provide reliable collision detection. We
never have a possibility of attempting an open on a directory.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>