Capture screen is not working well with CP enabled
The reason for the bug is that there is a race condition between regular
copy/paste and guest screen capture copy/paste.

1. There is image A on host clipboard.
2. After switch from host to guest, image A is copied to guest clipboard
3. User does a guest screen capture and it is put into host clipboard
4. After switch from guest to host, image A is copied again from guest
to host and overwrites the new guest screen capture in host clipboard.

The fix is not to do step 3. To do this, clipboard timestamp after step
2 is recorded. When user switches from guest to host (step 4), the
timestamp will be checked. If the timestamp is not changed, the guest
clipboard will not be copied to host.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>