lib/file: file locking upgrade
Improve the robustness and information logging of the process
validation used by the file locking code.

Define an improved, upwards compatible format for the execution
ID such that is is possible to include the name of the process
should it be available. This change is done such that older
code will run properly, even if it sees a newer lock file.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>