Fix OS X 10.7 unzip of files in shared folders using Archive Utility
Using Finder to double click on a zip file invokes the Archive
utility which fails to unzip a file on HGFS shares.
This was due to a read request of zero bytes being incorrectly handled.
The request was forwarded to the HGFS server which misinterpreted
the request for backdoor V3 read operation to instead using the
shared memory. This caused a protocol error to be returned from the
HGFS server.

The fix therefore is really two fold:
- The HGFS client should short-circuit the empty read and complete it
successfully immediately.
- The HGFS server should correctly handle zero length read requests
and not assume they will never be sent from any client.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>