lib/misc: Do not constrain the OS name string
The Hostinfo API for obtaining the OS name constrains the name
of the OS. These strings are highly variable in length and may
grow unexpectedly. Refactor the API to not care about the length
and fixup the call sites.

In a future change the internals of lib/misc will be overhauled
to also not care about the size. For now the (internal) size will
be inflated to more than large enough (512).

Another future change will remove any limitations found within the
tools code.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>