Eliminate duplicate handle field from VMCI resources.
Each VMCI resource has handle field, which is a
duplicate of the handle that is already stored within
the hash entry, which is also embedded inside the
resource.

This change removes the handle field and adds an
accessor to retrieve the handle from resource via
the hash entry.  The doorbell code has been modified
to use this accessor.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>