Use Linux ioctls for freezing filesystems.
The main purpose of this change is to allow using Linux ioctls that were
added in 2.6.29 (and backported to a few distro kernels) instead of our
own vmsync driver to freeze/thaw file systems.

This particular change required changing the way the POSIX sync driver
backend works, since it now has to try different backends to see what is
available or not. This means having a second "null" backend (the other
one is in the vmbackup plugin).

On top of that, I decided to clean up some of the code. Mainly:

. put the POSIX backends in their own files, make the syncDriverPosix.c
just call into the backends.

. get rid of the "DrivesAreFrozen()" call which wasn't really that useful,
and couldn't be implemented in some situations (e.g., the Linux ioctls).

. minor cleanups here and there.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>