[unity] Move glib signal from lib/unity to unity plugin.
The dnd plugin expects the unity plugin to emit a g_signal whenever
we enter or exit unity so that it can update some state. Since the
signal is strictly a plugin thing, move it out of the unity library
code and into the unity plugin code, which feels like a better fit.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>