shutdownStreamTest fails on RHEL6.1
The problem is that the test is broken for Linux, and it
appears to work with INET but actually it doesn't.

What happens with /st is that when we do the send(), we
actually get a SIG_PIPE, and the process exits, but there's
no error message, so it looks like the test passes.  Based
on this pass, when running the same test for vSockets, it
seems like we don't match the behavior because we fail.

But fixing the signal problem, by passing MSG_NOSIGNAL to
the send, makes it fail for INET too.  So now the test
fails for both system sockets and vSockets

The correct behavior on Linux is for the final recv() to
succeed.  Even with a local shutdown, the socket should
return 0 after a recv().

This change fixes the test to expect 0 on Linux (and
failure only for Windows, which is correct).

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>