Preserve hidden attribute for files copied from guest to MAC hosts.
Implemented the code to preserve the hidden attribute when a
file is copied from the gues to the MAC host.

Note: I have copied couple of functions from hgfsServerLinux.c. In
future, at some point of time, we should move those function to
a common file which will be consumed by various HGFS components.

Recently, we have submitted a changeset to preserve the permissions of
the directory when it is copied from the guest to host. As a part of
the changeset, a dummy fileInfo structure is added into the linked
list. When a user DND a folder from the guest to the host, we should
not display any progress when the dummy entry is processed. Modified
the code accordingly.

Modified the way, the arguments are passed while calling
HgfsSetHiddenXAttr in hgfsServerLinux.c. Re-factored the code
in ChangeInvisibleFlag() function.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>