Fix VMCI QP state transition on host detach
If the host detaches from a queue pair that is in an unmapped
state, it will erronously put it into a mapped state
(SHUTDOWN_MEM). This change makes the host detach move the
queue pair into SHUTDOWN_NO_MEM if the queue pair is already
in a *_NO_MEM state.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>