Fix asymmetric VMCI queue pair allocation
The entries in the VMCI queue pair broker are assumed to be
stored from the guest point of view, but if the queue pair is
created by the host endpoint, the sizes are not assigned
correctly. Instead, the produce queue will have the consume
queue size and vice versa. This change ensures that the values
stored in the queue pair broker entries are always from the
guest point of view.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>