VixTools: List Processes - Fix for potential core and couple of optimizations.
Change 1: Once we have found and printed the input pid we are looking for,
we can "break" the loop and move to the next input pid.

Change 2: The ignore for exited pid belongs to the outer "for" loop.

Change 3: Incorrect indexing into procList prints wrong process info. Also
may cause a core if "i >= procList->procCount".

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>