Fix dangling pointer in DynBufRealloc().
Coverity complains a lot about using a pointer
after it has been freed.

Indeed, for certain conditions, the buffer
is being freed, but a pointer to it is not being
set to NULL.

And since this function is actually trying to duplicate
the functionality of realloc(), simplify it a bit and
get rid of the offending piece of code.

From realloc() documentation:

"
In case that ptr is NULL, the function behaves exactly
as malloc, assigning a new block of size bytes
and returning a pointer to the beginning of it.

In case that the size is 0, the memory previously
allocated in ptr is deallocated as if a call to free
was made, and a NULL pointer is returned.
"

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>