lib/misc: centralize validating a PID
Code to determine is a PID is valid is tricky; we shouldn't have
multiple codes to do this. We happened to have a good, working code
for this in the file locking code. Move the code to lib/misc, adapt
it, slightly, to its new environment and then fix up the call sites.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>