Import dlpam_sterror symbol.
When the user executes any guest operation with invalid login
credentials, vmtoolsd process crashes [Signal 11, Segmentation fault].
Got the stack trace and found out that the crash happens in
Auth_AuthenticateUser() function. dlpam_strerror symbol is not loaded
from the PAM library but is referenced. This causes the crash.

Modified the code to import the dlpam_sterror symbol.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>