Make QP Broker locking local to vmciQueuePair.c
The locking for the VMCI queue pair broker was distributed
across the VMCI driver for the sole reason of being able to
handle a failed copy to user on Linux for queue pair
alloc. After this change, a failed copy to user will still
detach from an allocated queue pair, but a peer may have
attached to the queue pair between the alloc and detach. Since
the copy to user won't fail with a correctly working VMX, this
simplification is justified.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>