Fix QPair wakeup on suspend
On vmkernel, we do not unregister the guest memory, when a VM
is quiesced, so the queue pairs stay in a *_MEM state, if
already there. When detaching from a queue pair, we were only
waking up blocked kernel endpoints, if the queue pair was in
the *_NO_MEM state, since that would indicate that the QP is
quiesced. However, what we need to determine is whether the
queue pair headers are still mapped, since that is the common
operation for hosted and vmkernel on quiesce/unquiesce.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>