Lower (process, i/o) priority when wiping disks on Win32.
An attempt to avoid warnings popping up about applications that become
unresponsive because we're hogging resources by writing lots of stuff
to the disk.

Also be less chatty with the backdoor, and fix a bug where we were passing
the wrong parameters to the RPCI call making the VMX respond with a
failure.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>