vmbackup: couple of fixes to Linux freeze ioctl support.
. The sync driver code may get EIO and EACCES when trying to freeze
filesystems. The first happens on our HGFS mount, if HGFS is not
enabled on the host. The second, if a user has some filesystem
mounted with "700" permissions, so our code can't access it.
Just ignore those cases and keep going, freezing what we can.

. The error handler wasn't properly unfreezing frozen filesystems,
resulting in potentially unusable VMs if something went wrong.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>