Leak tls key 0
We have a couple of bugs that show system libraries on linux have a
nasty habit of destroying the first TLS key, even if they don't own it.
This is due to using uninitialized global variables (which default to
zero).  Rather than expose ourselves to the blame for such bugs,
let's just leak the key.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>