Implement separate HgfsSessions for Transport layer and Sessions layer.
To make the hgFileCopy more robust, we have decided to implement
separate sessions in the transport layer and the session layer. I
implemented a list of sessions in the transport layer. Moved the necessary
attributes from HgfsTransportSessionInfo to HgfsSessionInfo. Modified
all the necessary server connect|disconnect|receive|... functions to handle
the packets properly.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>