lib/lock: if you can't beat'em, join'em
Every time I change lib/lock to help find a bug it causes
problem if and only if I alter the size of a structure. This
indicates that there are multiple copies of lib/lock floating
around (known; as many as 5 copies have been observed), some that
may be incompatible with each other.

This change adds a real time signature and identification analysis
to lib/lock. It does this without changing any structure sizes. It
will help us catch the culprits and come up with a build plan to
handle this.

The run-time checking is done by having lib/lock always use a
unique signature, regardless of what type of build it is in. Any
lock that comes from a different copy of lib/lock will have a
different signature.

Previously, the signatures were static and also provided debugging
help to the type of lock.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>