lib/lock: finally put back the fix
This heavily reviewed change goes back in with one exception -
the part of the change that altered the size of structures isn't
included. The altering of structure sizes is causing hangs and
crashes. Hopefully the run-time checking, already in, will catch
the problems and we can clean those up. Once cleaned up we will
be free to alter structure sizes.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>