lib/lock: things go back
The entire new lib/lock structure is back in... except that
there is an ifdef for working the new way. By default it will
run compatible with the old way.

I've discovered that there are as many as 5 copies of lib/lock in
some of our products (i.e. Linux UI). Things would be fine if the
copies were used in a self contained fashion. When lib/lock runs
in the new way, random syndrome bits per library caused crashes
and hangs proving we're leaking locks across a boundary (something
unexpected).

Two copies of lib/lock passing around locks could be fatal if the
two copies are different versions of lib/lock.

Unfortunately logs and core dumps haven't been easy to find. I'm going
to begin a detailed search for the root cause via the ifdef; others
should not be affected.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>