Reduce heap allocations when listing processes in the guest.
Friday clean-up. We can just put the integer key in the data
structure that stores the state of the operation, and then pass the
address of that field to the hash table.
This does mean we need to use g_hash_table_replace() instead of
insert(), since we must update the key if we update the value.

Also, when passing the key along to timeout/clean-up function, just
stuff it into the pointer, since pointers are always large enough to
hold an int. This means more casts, but one fewer call to malloc().

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>