Implement 'InvalidateInactiveSessions' in hgfsServer.
This is the third part of the fix to make the HGFS server more robust.
Please check the wiki: https://wiki.eng.vmware.com/PGI/RobustVIXFileCopy for
more details.

Implmented a new function i.e. HgfsServerSessionInvalidateInactiveSessions
in HgfsServer. When this function is called, it checks all the available
HGFS server sessions and invalidate the inactive sessions. Everytime a
request is processed, session is tagged with timestamp when the request
is processed. If no request has been processed in the last 2 minutes, then
the session will be tagged as inactive and will be invalidated.

Callers can make use of this function invalidate the inactive session.
Implemented a watchdog in vixTools module which periodically sends this
request to the HGFS server.

Note: In this changeset, I have implemented the watchdog only in the
vixTools plugin. Not sure, if we require the same watchdog in another
modules.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>