Destory the sessions if they are inactive for a long amount of time.
This changeset is an extension to my earlier changeset #1592340. In
the changeset #1592340, I have implemented the 'InvalidateInactiveSessions'
function in the HGFS server. When this new function is called, it
checks for the inactivity of each session and invalidates the session
if it is inactive for a long amount of time. Also, we should ideally
implement another feature that destroys the session if the session is
inactive for a very long amount of time. Modified the code to destroy
the sessions besides just invalidating the sessions.

Also, implemented the code to limit the number of sessions that can be
created in one transport. HGFS server returns an error if the maximum
number of sessions are already created.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>