Modify the return value of HgfsServerSessionInvalidateInactiveSessions.
Modified the function signature for
HgfsServerSessionInvalidateInactiveSessions. With the new changes, the
function will return the number of active sessions left in the HGFS
server. This return value reflects the total number of sessions that still
need to be either invalidated or destroyed. Based on this return value,
the caller can make a decision whether to call the invalidator one more
time or not.

Modified the vixTools code to handle this new function signature.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>