Add VMCI RPC channel RIDs and allow messages to them from vSockets.
As part of the work to replace backdoor in the libRPC, we need two
additional VMCI channels in the VMX.  This change adds the resource IDs
for those channels.  One is privileged, meaning only admin/root guest
users can send to it.  The other allows messages from any user in the guest.

This change does not add handlers for the new channels, those will be added
in a future change.

This change also modifies vSockets to allow DGRAM sockets to send messages
over these new channels (we have explicit checks when sending to the VMX).

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>