Remove the special logic implemented for default session.
Recently, I have submitted a changeset #1571619 to make the HGFS
server more robust for the file transfers. As a part of that
changeset, we have implemented the separate HGFS sessions for
Transport layer and Sessions layer. In order to handle old file
copy library clients, we have implemented the concept of default
session. This default session is created when the transport
session is initially created and exists till the transport sessions
is destroyed.

But, because of this special default session, we had implemented
special logic in various parts of the code. In future, it will be
difficult to maintain/write any new code. We decided to remove
this special logic. Modified the HGFS server code to get rid of the
special handling for default session.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>