Execute HGFS_ASSERT_INPUT only in successful code paths.
Check bug 816476 for more details. In changeset #1611812, we have
removed the special logic related to the default session. So, in case of
any error, HgfsParseRequest() doesn't guarantee that all the attributes of
HgfsInputParam datastructure will be filled. In case of an error, the
code should not execute HGFS_ASSERT_INPUT.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>