lib/lock: MXUser statistics are too expensive
Dmitriy specifically wanted the statistics measurement code to
actually measure the time it took to acquire an uncontended lock.
This turns out to be a very large overhead and not something we
really care about or can measure accurately.

Refactor the MXUser statistics code to only measure the acquisition
time when the lock is contended - most of our locks will have virtually
no overhead in a statistics build this way.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>