Return an error in 'mount' call when shared folders are disabled
WHile adapting to the newer kernel API we have changed the way a root
dentry is filled. At the time of mount call, we don't send any request
to the HGFS server to check if the shared folders are enabled or not.
So, irrespective of the state of the HGFS server in the VMX, the 'mount'
call succeeds. We should ideally return an error if the shared folders
are disabled.

This is done by issuing a getAttr request to the server and returning
an error if the getAttr request fails.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>