lib/lock: No need for stats all of the time
Way back when there was a concern that a release build might need
to be performance analyzed (for locks) in the field. At the time the
VMX wasn't considered to be a major bottleneck so the MXUser
statistics were built into every type of build.

Now a days, we have stats builds in the field when we need numbers
and the VMX is turning into a bottleneck. Remove the stats code
from all but stats builds. This saves ~7 clocks when acquiring
an uncontended exclusive lock in a release build.

The statsMem pointer is left in all builds to maintain the size of
the structure until will can find where a lock is used across two
copies of lib/lock with different build types.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>