Fix HGFS Server malformed header for change notification events sent to the clients.
The change notification events generate HgfsRequestNotifyV4 requests
to be sent to the clients to report changes. Unfortunately, the
server packed the header badly as it calculated the total packet
size incorrectly: the size of file name length field was added twice.

This means that the clients that handle notification events will
receive them prefixed with headers with	the packetSize field greater
than the total packet size sent.

This was not noticed until now as the Windows HGFS client (currently
the only client enabled for change notification support) did not
validate the header from the server request correcty or thoroughly.
When this code was corrected, the notification requests failed with
a protocol error validating the header.

Furthermore, since this code has shipped in the HGFS server,
clients have to effectively ignore that field for these request
type.

Cleaned up some code, so that unused functions now used instead
of duplicating code.

Cleaned up some possible memory leaks too.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>