Provide GuestOS functionality for the scheme code.
This attempts to provide the functionality used by monitor mode that
relates to guest OS types.  Doing this required re-jiggering some of
the enums and tables used by the guest_os so that we could get a hold
of them from the scheme code.  Mostly this involved making macro
generators that can then be instantiated in more than one place.  With
that in place the actual scheme implementation was pretty
straightforward.

One small unfortunate detail is that the strings used by the guestOS
code are often generated by letting the compiler concatenate two
strings together, but when using macros to pull this into scheme we
end up with two distinct strings.  This meant the scheme code had to
include a string-append to fix this for every entry even though most
didn't need it.  As one small bit of premature optimiztion, I modified
the implementation of string-append to special case the case of having
only one string.

I got scared away from table-ifying the strings themselves due to the
comments about GuestInfoGetOSName.

I also modified VMOvhdGetVMConfigStr slightly so that it can be passed
a NULL (or nil) default string.

Testing:
Played around with the new scheme functions and some logging.
Used some logging and some VMs from eregsuite to make sure the C version
of GuestOSID didn't give different results.
frobos (BT-intel, BT-amd, VT, SVM, EPT, RVI)

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>