Snapshot's file absolute path must be respected
It is just wrong to take an absolute pathname, truncate it to a basename,
then do a path search on the basename to find the file. It seems incredible
that such a blunder has existed in a common routine since 2004.

Made path search semantics more sensible: only unadorned basename
is searched via path.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>