No need to lock twice before accessing VMCI context queue pair array.
The VMCI context keeps track of queue pairs allocated by that
context in a handle array. This handle array is always
accessed while holding the general queue pair lock, so we do
not need to grab the context lock as well. In fact, the header
file already describes this. So drop grabbing the context lock
for queuePairArray.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>