Change File_CreateDirectoryHierarchy to enable "undo"
When calling File_CreateDirectoryHierarchy, there was no way to know
what was actually created by the call and what wasn't. This meant that
trying to be a good neighbor and clean up after ourselves if whatever we
wanted the directory for failed (or was cancelled) wasn't possible.

This change adds an optional OUT parameter to that gets set to the
top-most directory which was actually created by the call. If the caller
then wants to clean up, they can pass that into
File_DeleteDirectoryTree.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>