Fix callback in vmcilib
There's a nasty bug where, if the detach callback gets invoked at the
same time that the client is releasing, we can end up executing after
the client has unloaded.  The problem is that the callback will wait on
the lock, and once Release() lets it go, the client is allowed to unload.
The lock, and the rest of the code for the callback, is then gone.

The fix is two-fold:

1. Use the system callback mechanism.  This utilizes callback objects
stored in the system object table.  The callback object is shared between
the publisher and the subscriber using a symbolic name.  This provides
a level of indirection between VMCI and vmcilib, i.e., they don't have
to rely on each other being around.

2. Introduce a new function, VMCI_Exit().  The client calls this function
when it unloads, at PASSIVE level, with no locks held, and this
unregisters from the callback.  So the new workflow for a client is as
follows:

VMCI_DeviceGet()
VMCI_DeviceRelease()
VMCI_DeviceGet()
VMCI_DeviceRelease()
...
VMCI_Exit()

The above changes require no modification to the VMCI/vmcilib IOCTLs and
so are backwards-compatible with shipping clients that do not use
callbacks (we haven't yet shipped any that do).

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>