Improper ASSERTs in vmx cpuid handling
2 ASSERTs can trigger due to input values from the guest cpuid call. The
first one happens if the input level falls within the valid low range
level, but hits a hole where no level exists. The second ASSERT triggers
if the ecx value is out of range for the TOPOLOGY level on Intel.  The
fix is to handle the values at runtime and not assert.  Also fix a
couple of line wraps.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>