Introducing new XML tag <cmd>...</cmd> for command name in ListProcessesInGuest.
This is to relieve VIX Client code from having to parse OS specific
command names from the full command line, except for Windows, which
has been left as it is. Without this fix, current implementation on
Mac, Linux, Solaris and FreeBSD are broken whenever we have folders
or executable file names with blanks spaces in them.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>