Call vmwarectrl to set host GUI topology even if we use RandR12
The new vmwgfx driver needs to be separately informed about the host GUI
topology since with screen objects we differentiate between that and
Guest FB layout, which is set using RandR12.

We could in theory set the host gui layout using direct kernel calls to
the modesetting system of the kernel, but we won't do that until we've
sorted out the security implications.

To be paranoid, we grab the X server around the host GUI information
update and the guest FB layout update, taking care to really flush the
XUngrabServer call.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>