HGFS: fix OOPS in d_alloc_name() which can't take NULL anymore
During the intial hgfs mount process we need a root dentry. We can get
the reference to the root dentry by callig d_alloc_root(). But we cannot
call d_alloc_root in HgfsReadSuper() since it requires a valid inode
which we don't have at that point of time. We implemented a workaround
for this issue. The workaround was done in two phases as specified below:
- Create a dummy root dentry by calling d_alloc_name().
- Send a getAttr request to the hgfs server and get the root inode.
- Populate the content of the dummy root entry with the root inode.

Unfortunately in newer kernel versions the semantics of d_alloc_name()
is changed. We can no longer pass NULL to the d_alloc_name (which we
have been doing till now). Fixed this by implementing a new function
HgfsGetRootDentry() that first gets the root inode, fills the properties
of the root inode and then uses d_alloc_root() to generate the root
dentry.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>