Fix File_PathJoin ASSERTs some more
I think that File_PathJoin's ASSERTs were a bit too aggressive.  I
believe that the intention was to disallow UNC paths for the tail
portion of the path, not to disallow all paths that start with a
directory separator. (Allowing a leading directory separator on Linux
but not on Windows seems a bit dangerous anyway.)

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>