GOS: is it an enum or not
Some time ago the GOS table ID values were changed from defines to
an enum however the "enumity" of the values was not fully plumbed
through the source base. Fix this so we get all of the type checking
and visual goodness.

Did a little bit of grouping/rearranging to the guest ID list to make
it clear that the order can be changed (it's for internal use only).
This will make it more obvious how many of each OS we have and make
searching for things that can be compacted easier.

This is the first of a series of changes to create some room for
more guest OSen - we're limited to 64 as things stand (and they
are all used).

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>