Audit and remove VMX86_WGS
wgs/gsx/server (non-esx) code still lives within its older still active
siblings. I located as much as possible and removed its dependencies and
tried to make appropriate changes related to the code being removed.
This change had the potential to remove a lot of other old products and
tools, but I resisted that temptation (console, mui, ccagent to name a
few). For the most part it was a standard removal.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>