Make the first of a few changes to accomodate Haswell.
The new Haswell CPU seems to behave like the other newish Intel
architectures, so in the main the changes needed are simple. I've added
the  macros to allow for detection for the model family and
architecture. The interpreter changes are to make the interpreted
results consistent with native execution. I also added Haswell to the
list of architectures supported by vpmc. The test changes are the
counterpoint to the monitor changes.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>