Add NULL checks before accessing logger attribute.
When we login into a VM, a new VMware Tools process starts with the
logged in User account.  If the logged in user is the least prviliged
user (guest account in Windows 8), then the process may not be able to
create a logger handler. GlibUtils_Create*Logger() functions may return
NULL.  In such case, accessing the logger handler may result in a crash.

Modified the code to add proper NULL checks everywhere before
accessing the logger handler.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>