DescriptorWrite: revert to in-place write on NFS
The ESX NFSv3 client can't handle rename. So we need to switch back to
the write-the-descriptor codepath on NFS on ESX.

Note that even though we don't use the result of the AtomicTempFile
creation, failures to create/write the AtomicTempFile are still
treated as fatal (see descriptor.c: errorWriteDirect
HostType_OSIsVMK); while there are a few scenarios where we might
be able to continue despite them, the storage is fatally sick
and we should error early rather than limp along.  Furthermore,
special casing these failure paths requires significant additional
code (and therefore development work and risk) with no benefit.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>