Add support for promoted AVX2 instructions
This changeset adds support for AVX instructions promoted to enable
use of the full 256bits of the ymm regs by the AVX2 feature set. This
changeset does not introduce any new instructions or encodings. It
just honors the VEX prefix 'L' bit used to indicate the width of the
source and destination operands. Previously, these instructions
would result in a #UD fault under AVX if VEX.L==1.

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>