vPageChannel: Failed to allocate packet (channel=ffff880013fda380) (size=69600)
We're trying to perform really large atomic allocations, for example
when allocating a packet to send receive buffers down to the host.  Such
allocations can be hard to fulfil when the kernel cannot block.  So make
the chunks even smaller when we're in an atomic context.

One potential downside to this is that we will signal a lot, i.e., we're
going to signal 16x more for page-sized than for DGRAM-sized packets.  We
fix this by writing out all of the packets and then only signalling
afterwards.  The peer is okay with this: it may peek a packet header but
then not see sufficient data in the QP to be able to read the rest, so it
will bail out, leaving the packet for when all the sends are complete and
it gets signalled.

The other, and more proper, way to fix this is to change the interface so
that packets are iovecs of page-sized chunks.  That's a somewhat more complex
change that we aim to do for v2 anyway (actually we'll have packet lists,
and each packet in the list will be split into page-sized chunks, and each
chunk will hold actual PPNs).

Signed-off-by: Dmitry Torokhov <dtor@vmware.com>